\name{BenthicLatent-package}
\alias{BenthicLatent-package}
\alias{BenthicLatent}
\docType{package}
\title{
\packageTitle{BenthicLatent}
}
\description{
\packageDescription{BenthicLatent}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BenthicLatent}
\packageIndices{BenthicLatent}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BenthicLatent}

Maintainer: \packageMaintainer{BenthicLatent}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
